Refactor fit() into training_tensors.ts
